+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[unreleased]
+
+* Add debug logs.
+
 [2.0.3-alpha.14]
 
 * Add community metadata validation.